need to repair collecting data from additional car info

+++ src/main/java/core/calculator.java
@@ -19,7 +19,7 @@ public class calculator {
     public static String carGeneration = null; // generation of the car I, II, III
     public static boolean gasDesiel; // false - diesel, true - gas
     public static float engineCapacity; // engine capacity - for example 1.1, 2.0, 3.0, 2.8
-    public static byte amountOfDoors; // amount of door in the car - 3,5
+    public static int amountOfDoors; // amount of door in the car - 3,5
     public static int counterStatus; // current counter status in the car
     public static int kmInYear; // how many kilometers the client will drive in a year
     public static String wherePark = null; // where the client will park his/her car (outdoor, garage)
